// packages/shared/types/database.d.ts
export type Json = string | number | boolean | null | {
    [key: string]: Json | undefined;
} | Json[];
export type Database = {
    __InternalSupabase: {
        PostgrestVersion: "14.1";
    };
    public: {
        Tables: {
            accounts: {
                Row: {
                    account_type: string;
                    admin_id: string;
                    business_type: string | null;
                    country: string;
                    created_at: string;
                    currency: string | null;
                    data: Json;
                    deleted_at: string | null;
                    email: string | null;
                    id: string;
                    is_charges_enabled: boolean;
                    is_details_submitted: boolean;
                    is_payouts_enabled: boolean;
                    platform_fee_fixed: number | null;
                    platform_fee_percent: number | null;
                    processor_fees_applicable: boolean;
                    status: string;
                    stripe_id: string | null;
                    updated_at: string;
                };
// ...
                Update: {
                    account_type?: string;
                    admin_id?: string;
                    business_type?: string | null;
                    country?: string;
                    created_at?: string;
                    currency?: string | null;
                    data?: Json;
                    deleted_at?: string | null;
                    email?: string | null;
                    id?: string;
                    is_charges_enabled?: boolean;
                    is_details_submitted?: boolean;
                    is_payouts_enabled?: boolean;
                    platform_fee_fixed?: number | null;
                    platform_fee_percent?: number | null;
                    processor_fees_applicable?: boolean;
                    status?: string;
                    stripe_id?: string | null;
                    updated_at?: string;
                };
                Relationships: [
                    {
                        foreignKeyName: "accounts_admin_id_fkey";
                        columns: ["admin_id"];
                        isOneToOne: false;
                        referencedRelation: "users";
                        referencedColumns: ["id"];
                    }
                ];
            };
// ...
        };
// ...
    };
};